pushing solutions for exercises 15-23 from part 4

+++ part4/blog_list/controllers/blogList.js
@@ -1,16 +1,53 @@
 const blogsRouter = require('express').Router();
+const jwt = require('jsonwebtoken');
 const Blog = require('../models/blogs');
+const User = require('../models/users');
+const { SECRET } = require('../utils/config');
+const { userExtractor } = require('../utils/middleware');
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 });
   response.json(blogs);
 });
 
-blogsRouter.post('/', async (request, response, next) => {
-  const blog = new Blog(request.body);
+blogsRouter.get('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id).populate('user', {
+    username: 1,
+    name: 1,
+  });
+
+  if (blog) {
+    response.json(blog.toJSON());
+  } else {
+    response.status(404).end();
+  }
+});
+
+blogsRouter.post('/', userExtractor, async (request, response, next) => {
+  const { url, title, author, likes } = request.body;
+  const user = request.user;
+  const token = request.token;
 
   try {
+    const decodedToken = jwt.verify(token, SECRET);
+
+    if (!(token && decodedToken.id)) {
+      next();
+    }
+
+    const blog = new Blog({
+      url,
+      title,
+      author,
+      likes,
+      user: user.id,
+    });
+
     const savedBlog = await blog.save();
+
+    user.blogs = user.blogs.concat(savedBlog._id);
+
+    await user.save();
 
     response.status(201).json(savedBlog);
   } catch (exception) {
@@ -18,8 +55,29 @@ blogsRouter.post('/', async (request, response, next) => {
   }
 });
 
-blogsRouter.delete('/:id', async (request, response, next) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response, next) => {
+  const user = request.user;
+  const token = request.token;
+
   try {
+    const decodedToken = jwt.verify(token, SECRET);
+
+    if (!(token && decodedToken.id)) {
+      next();
+    }
+
+    const { user: blogUser } = await Blog.findById(request.params.id);
+
+    // not sure why this happens but I have to convert the user entry
+    // to a string in here instead of the model file.
+    // Both the toJSON and toObject methods were tried, but if
+    // blog.user is destructured it always returns as an Object
+    // instead of a String
+    if (blogUser.toString() !== user.id) {
+      return response
+        .status(401)
+        .send({ error: "this user can't delete this blog" });
+    }
     await Blog.findByIdAndRemove(request.params.id);
     response.status(204).end();
   } catch (exception) {
